requirements.txt
# Core Data Processing
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2

# Visualization
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0

# Web Framework
streamlit==1.28.2
flask==3.0.0

# AI Integration
openai==1.3.7
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
xlsxwriter==3.1.9

# Testing
pytest==7.4.3
pytest-cov==4.1.0
